+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Create via menu doesn't create as child of selected object
 - Bugfix: Drift in FreeLook with SimpleFollow
 - Bugfix: Post-processing profiles not cleaned up when no active vcams
+- New sample scene demonstrating new CinemachineConfiner2D extension.
 
 
 ## [2.7.1] - 2020-11-14
