+++ Samples~/Cinemachine Example Scenes/Shared/Scripts/SimpleElevator.cs.meta
